# src/api/openapi.yml
openapi: 3.0.0
info:
  version: 0.1.0
  title: AgentENV API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
    # AdminApiKeyAuth / AdminTeamAuth: alphabetical names ensure token
    # validation runs before team context population.
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    sandboxID:
      name: sandboxID
      in: path
      required: true
      schema:
        type: string
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    CPUCount:
      type: integer
      format: int32
      minimum: 1
      description: CPU cores for the sandbox
    MemoryMB:
      type: integer
      format: int32
      minimum: 128
      description: Memory for the sandbox in MiB
    DiskSizeMB:
      type: integer
      format: int32
      minimum: 0
      description: Disk size for the sandbox in MiB
    DrainRequest:
      type: object
      properties:
        deadlineMs:
          type: integer
          format: uint64
          description: >-
            How long this pass may spend pausing and publishing. Omitted, the
            node's own configured deadline applies. A pass reports what is left
            rather than running to completion, so the caller polls.
    DrainProgress:
      type: object
      required:
        - remaining
        - published
        - failed
      properties:
        remaining:
          type: integer
          format: uint32
          description: >-
            Sandboxes on this node that are not paused yet. Zero means nothing
            is lost by stopping the process.
        published:
          type: integer
          format: uint32
          description: >-
            Paused sandboxes this pass committed to the repository, so another
            node can take them over.
        failed:
          type: integer
          format: uint32
          description: Sandboxes this pass tried and could not pause or publish.
...
    SnapshotInfo:
      type: object
      required:
        - snapshotID
        - names
        - cpuCount
        - memoryMB
        - diskSizeMB
        - createdAt
        - updatedAt
      properties:
        snapshotID:
          type: string
          description: The actual snapshot ID (stable identifier). Always contains the raw snapshot UUID, never an alias.
        names:
          type: array
          items:
            type: string
          description: User-provided aliases for the snapshot. Empty if no alias was assigned during snapshot creation.
        cpuCount:
          $ref: "#/components/schemas/CPUCount"
        memoryMB:
          $ref: "#/components/schemas/MemoryMB"
        diskSizeMB:
          $ref: "#/components/schemas/DiskSizeMB"
        createdAt:
          type: string
          format: date-time
          description: Time when the snapshot was created
        updatedAt:
          type: string
          format: date-time
          description: Time when the snapshot was last updated
        imageRef:
          type: string
          description: >
            OverlayBD-native OCI image reference for the snapshot rootfs when
            source-registry image publication was enabled and succeeded.
            Omitted when no rootfs image was published.
...
    SandboxSnapshotRequest:
      type: object
      properties:
        name:
          type: string
          description: Optional name for the snapshot template. If a snapshot template with this name already exists, a new build will be assigned to the existing template instead of creating a new one.
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error
...
paths:
...
  /sandboxes/{sandboxID}/snapshots:
    post:
      summary: Create snapshot
      description: Create a persistent snapshot from the sandbox's current state. Snapshots can be used to create new sandboxes and persist beyond the original sandbox's lifetime.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxSnapshotRequest"
      responses:
        "201":
          description: Snapshot created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotInfo"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
...
  /admin/drain:
    post:
      summary: Drain the node
      description: >-
        Stops this node admitting new sandboxes, then pauses and publishes what
        it is already holding for as long as the deadline allows. Irreversible
        for the life of the process, and bounded rather than complete: one call
        reports what is left, so a preStop hook polls it until `remaining` is
        zero.
      tags: [admin]
      security:
        - AdminApiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrainRequest"
      responses:
        "200":
          description: The drain pass completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrainProgress"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"